+++ code/includes/ServerConfiguration.hpp
@@ -6,28 +6,21 @@
 
 # include "Route.hpp"
 
-# define ERROR_404_STR "/custom_404.html"
-# define ERROR_404_INT 404
-# define ERROR_405_STR "/custom_405.html"
-# define ERROR_405_INT 405
-# define ERROR_500_STR "/custom_500.html"
-# define ERROR_500_INT 500
-
 /// @brief The configuration specific for each server
 class ServerConfiguration
 {
 public :
 
-	ServerConfiguration(uint32_t host, uint16_t port, std::vector<std::string> serverNames, \
-	std::map<unsigned short, std::string> errorPages, size_t maxClientBodySize, \
-	std::map<std::string, Route> routes, std::string root);
+	ServerConfiguration(std::vector<std::string> serverNames, \
+						std::map<unsigned short, std::string> \
+						errorPages, size_t maxClientBodySize, \
+						std::map<std::string, Route> routes, \
+						std::string root);
 	ServerConfiguration(ServerConfiguration const &src);
     ServerConfiguration    &operator=(ServerConfiguration const &src);
 	~ServerConfiguration(void);
 
 	const std::string							&getRoot(void) const;
-	const uint32_t								&getHost(void) const;
-	const uint16_t								&getPort(void) const;
 	const std::vector<std::string>				&getServerNames(void) const;
 	const std::string							&getErrorPage(unsigned short error) const;
 	const std::map<unsigned short, std::string>	&getErrorPages(void) const;
@@ -36,8 +29,6 @@ public :
 	const Route									*getOneRoutes(std::string path) const;
 
 private :
-	uint32_t								host;
-	uint16_t								port;
 	std::vector<std::string>				serverNames;
 	/// @brief ushort : error code, std::string, page path
 	/// Associate an error with a page, returned to the client.
